Inline assigned-to filter wrapper; keep Unassigned on member pick
Merge the passthrough wrapper and its inner into one component (the {#key}
was already removed for live filtering) and delete the now-redundant file.
Set inputText on a member pick so "Unassigned" stays visible for a selected
member.
